Send broken submission emails to committers, even to people who are not registered as users of CDash. But only if a project admin actively sets up the build group settings to do so.
Added 'emailcommitters' field to the buildgroup database table.
Added corresponding GUI checkboxes to the manageBuildGroup.php page.
This new setting defaults to 0, so it does *not* email non-registered-user
committers by default.

Implemented changes to the CDash sendemail function to send emails out to
committers when the new build group setting 'emailcommitters' is on. These
emails are sent even to committers who are *not* registered users of CDash.

Modified the 'actualtrilinossubmission' test to set the emailcommitters
field to 1. Manually verified that the log file contains text indicating the
expected emails have been sent out. We still need to add strong automated
checking of this test result.

Added CDASH_WARN_ABOUT_UNREGISTERED_COMMITTERS config variable, to enable
squelching the stupid warning everybody always gets about unregistered users
submitting code to their project. By default it's set to 1 to keep the
behavior the same as before, emitting the stupid warning, but if you would
like, you may now configure your CDash server, setting it to 0. And you'll
never have to see that stupid warning again.

Added CDASH_TESTING_RENAME_LOGS config variable and corresponding CMake
option to enable keeping the log files, so they are available for inspection
after running the test suite. By default, this is OFF/false to keep the
behavior the same as before.

Added code to tests that inject config settings so that later readers of the
config file can see where they came from...

+++ tests/kwtest/kw_web_tester.php
@@ -102,15 +102,19 @@ class KWWebTestCase extends WebTestCase {
     {
     if(file_exists($filename))
       {
-      // Delete file:
-      unlink($filename);
-
-      // Alternatively, rename it to a random name to keep for later
-      // inspection. (Comment out the above unlink and uncomment this
-      // chunk to keep the log files from the test suite around...)
-      //
-      //global $CDASH_LOG_FILE;
-      //rename($filename, $CDASH_LOG_FILE . "." . mt_rand() . ".txt");
+      global $CDASH_TESTING_RENAME_LOGS;
+      if ($CDASH_TESTING_RENAME_LOGS)
+        {
+        // Rename to a random name to keep for later inspection:
+        //
+        global $CDASH_LOG_FILE;
+        rename($filename, $CDASH_LOG_FILE . "." . mt_rand() . ".txt");
+        }
+      else
+        {
+        // Delete file:
+        unlink($filename);
+        }
       }
     }
